+++ spec/controllers/staff/organistations_controller_spec.rb
@@ -1,0 +1,12 @@
+require "rails_helper"
+
+RSpec.describe Staff::OrganisationsController do
+  context "when the user is not logged in" do
+    it "redirects the user to the root path" do
+      get :index
+      expect(response).to redirect_to(root_path)
+      get :show, params: {id: double}
+      expect(response).to redirect_to(root_path)
+    end
+  end
+end