+++ spec/features/staff/users_can_view_transactions_on_an_activity_page_spec.rb
@@ -14,9 +14,7 @@ RSpec.feature "Users can view transactions on an activity page" do
         transaction = create(:transaction, parent_activity: activity, value: "100")
         other_transaction = create(:transaction, parent_activity: other_activity, value: "200")
 
-        visit activities_path
-
-        click_link activity.title
+        visit organisation_activity_path(activity.organisation, activity)
 
         expect(page).to have_content(transaction.value)
         expect(page).to_not have_content(other_transaction.value)
@@ -26,9 +24,7 @@ RSpec.feature "Users can view transactions on an activity page" do
         transaction = create(:transaction, parent_activity: activity)
         transaction_presenter = TransactionPresenter.new(transaction)
 
-        visit activities_path
-
-        click_link activity.title
+        visit organisation_activity_path(activity.organisation, activity)
 
         expect(page).to have_content(transaction_presenter.financial_quarter_and_year)
         expect(page).to have_content(transaction_presenter.value)
@@ -39,9 +35,7 @@ RSpec.feature "Users can view transactions on an activity page" do
         transaction_1 = create(:transaction, parent_activity: activity, date: Date.today)
         transaction_2 = create(:transaction, parent_activity: activity, date: Date.yesterday)
 
-        visit activities_path
-
-        click_link activity.title
+        visit organisation_activity_path(activity.organisation, activity)
 
         expect(page.find("table.transactions tbody tr:first-child")[:id]).to eq(transaction_1.id)
         expect(page.find("table.transactions tbody tr:last-child")[:id]).to eq(transaction_2.id)
@@ -58,13 +52,7 @@ RSpec.feature "Users can view transactions on an activity page" do
         transaction = create(:transaction, parent_activity: project_activity)
         transaction_presenter = TransactionPresenter.new(transaction)
 
-        visit activities_path
-
-        click_link fund_activity.title
-        click_on t("tabs.activity.children")
-        click_link programme_activity.title
-        click_on t("tabs.activity.children")
-        click_link project_activity.title
+        visit organisation_activity_path(project_activity.organisation, project_activity)
 
         expect(page).to have_content(transaction_presenter.financial_quarter_and_year)
         expect(page).to have_content(transaction_presenter.value)
@@ -74,13 +62,7 @@ RSpec.feature "Users can view transactions on an activity page" do
       scenario "transactions cannot be created or edited by a BEIS user" do
         transaction = create(:transaction, parent_activity: project_activity)
 
-        visit activities_path
-
-        click_link fund_activity.title
-        click_on t("tabs.activity.children")
-        click_link programme_activity.title
-        click_on t("tabs.activity.children")
-        click_link project_activity.title
+        visit organisation_activity_path(project_activity.organisation, project_activity)
 
         expect(page).to_not have_content(t("page_content.transactions.button.create"))
         within("tr##{transaction.id}") do